refactor layout component
