+++ src/components/Schedule/style.sass
@@ -1,9 +1,3 @@
-.container-schedule
-    width: 100vw
-    height: 100vh
-    background-color: #d6d6d6
-    font-family: 'Montserrat', sans-serif
-
 .schedule-title
     font-size: 2rem
     font-weight: 700
