+++ src/components/SectionStyle.sass
@@ -1,0 +1,27 @@
+%container-light
+    width: 100vw
+    min-height: 100vh
+    background-color: #dfdfdf
+    font-family: 'Montserrat', sans-serif
+
+%container-dark
+    width: 100vw
+    min-height: 100vh
+    background-color: #444040
+    font-family: 'Montserrat', sans-serif
+    
+.container-who 
+    @extend %container-light
+
+.container-what
+    @extend %container-dark
+
+.container-schedule
+    @extend %container-light
+
+.container-contact
+    @extend %container-dark
+    
+
+.animated
+    height: 100%
