+++ src/components/WhatWeDo/index.js
@@ -1,50 +1,33 @@
-import React, { useEffect, useState } from 'react'
-import { I18NConsumer } from '../i18ncontext'
-import { Animated } from 'react-animated-css'
-import isInViewPort from '../../utils/isInViewport'
-import Content from './content'
+import React, { useState, useEffect } from 'react'
 
 import './style.sass'
 
-
 export default function (props) {
-    const [isInView, setIsInView] = useState(false)
+    const [widthState, setWidthState] = useState({ width: "10%" })
+    const { text, lang } = props
 
     useEffect(() => {
-        function watchScroll() {
-            window.addEventListener('scroll', function (event) {
-                if (isInViewPort('#what')) setIsInView(true)
-            })
-        }
+        setWidthState({ width: "100%" })
+    }, [])
 
-        watchScroll()
-        //remove listener (like componentWillUnmount)
-        return () => {
-            window.removeEventListener("scroll", function (event) {
-                isInViewPort('#what')
-            })
-        }
-    })
+    return <>
+
+        <div style={{
+            backgroundColor: "#4577f7",
+            width: widthState.width,
+            transition: "width 1s linear"
+        }}>
+            <p className="what-title section-title-wrapper">{text[`${lang}`].navbar.what}</p>
+        </div>
+
+        <div>
+            <article>
+                <p>Lorem ipsum dolor sit amet consectetur adipiscing elit euismod proin, mauris himenaeos praesent posuere vehicula nascetur ante interdum velit pellentesque, neque nam nulla sodales sed felis cum molestie. Metus ullamcorper tortor magna commodo mauris enim suscipit consequat, bibendum nec nascetur lectus praesent facilisi. Montes ornare ad orci lobortis bibendum, neque diam eget scelerisque cursus condimentum, donec semper mauris accumsan.</p>
+            </article>
+        </div>
 
 
-    return (
-        <>
-            <I18NConsumer>
-                {({ text, lang }) => (
-                    <div className="container-what" id="what">
-                        {isInView && <>
-                            <Animated
-                                animationIn="fadeIn"
-                                animationInDelay={300}
-                                isVisible={true}>
-                                <Content text={text} lang={lang} />
-                            </Animated>
-                        </>}
-                    </div>
-                )}
-            </I18NConsumer>
-        </>
-    )
 
+
+    </>
 }
-
