+++ src/components/WhatWeDo/style.sass
@@ -1,9 +1,3 @@
-.container-what
-    width: 100vw
-    height: 100vh
-    background-color: #444040
-    font-family: 'Montserrat', sans-serif
-
 .what-title
      font-size: 2rem
      font-weight: 700
