+++ src/components/WhoAreWe/style.sass
@@ -1,12 +1,49 @@
-.container-who
-    width: 100vw
-    height: 100vh
-    background-color: #dfdfdf
-    font-family: 'Montserrat', sans-serif
-    
 .who-title
      font-size: 2rem
      font-weight: 700
      color: white
+
+.who-container
+     margin: 1rem 2rem
+
+     &__article-intro
+          margin: 0 3rem
+          padding: 0.3rem 1rem
+          background-color: #f5f5f5
+          border-radius: 10px
+
+.pictures-container
+     display: flex
+     width: 100%
+
+.pic-wrapper
+     margin-top: 1rem
+     width: 100%
+
+.who-pictures
+     display: flex
+     justify-content: center
+
+
+.coach1-pic
+     width: 45%
+     transition: width .5s ease
+
+.coach1-pic:hover
+     width: 50%
+
+.coach2-pic
+     width: 45%
+     transition: .5s ease
+     background-color: rgba(255, 255, 255, 0.6)
+     border-radius: 50%
+
+
+.coach2-pic:hover
+     width: 50%
+    
+
+
+
      
-     
+
