+++ src/index.sass
@@ -9,3 +9,6 @@ ul
 
 .section-title-wrapper
     margin: 0 2rem
+
+.section-intro
+    font-size: 1.3rem
