+++ src/text.js
@@ -13,7 +13,10 @@ const text = {
             schedule: "Schedule",
             contact: "Contact"
         },
-        header: <p className="quote">BE YOUR<br />BETTER<br />VERSION</p>
+        header: <p className="quote">BE YOUR<br />BETTER<br />VERSION</p>,
+        sectionWho: {
+            intro: "Lorem ipsum dolor sit amet consectetur adipiscing elit euismod proin, mauris himenaeos praesent posuere vehicula nascetur ante interdum velit pellentesque, neque nam nulla sodales sed felis cum molestie. Metus ullamcorper tortor magna commodo mauris enim suscipit consequat, bibendum nec nascetur lectus praesent facilisi. Montes ornare ad orci lobortis bibendum, neque diam eget scelerisque cursus condimentum, donec semper mauris accumsan."
+        }
 
     },
 
@@ -24,7 +27,11 @@ const text = {
             schedule: 'Horario',
             contact: 'Contacto'
         },
-        header: <p className="quote">SE TU<br />MEJOR<br />VERSIÓN</p>
+        header: <p className="quote">SE TU<br />MEJOR<br />VERSIÓN</p>,
+        sectionWho: {
+            intro: "Lorem ipsum dolor sit amet consectetur adipiscing elit euismod proin, mauris himenaeos praesent posuere vehicula nascetur ante interdum velit pellentesque, neque nam nulla sodales sed felis cum molestie. Metus ullamcorper tortor magna commodo mauris enim suscipit consequat, bibendum nec nascetur lectus praesent facilisi. Montes ornare ad orci lobortis bibendum, neque diam eget scelerisque cursus condimentum, donec semper mauris accumsan."
+        }
+
     },
 
     CAT: {
@@ -34,7 +41,10 @@ const text = {
             schedule: 'Horari',
             contact: "Contacte"
         },
-        header: <p className="quote">SIGUES<br />LA TEVA<br />MILLOR<br />VERSIÓ</p>
+        header: <p className="quote">SIGUES<br />LA TEVA<br />MILLOR<br />VERSIÓ</p>,
+        sectionWho: {
+            intro: "Lorem ipsum dolor sit amet consectetur adipiscing elit euismod proin, mauris himenaeos praesent posuere vehicula nascetur ante interdum velit pellentesque, neque nam nulla sodales sed felis cum molestie. Metus ullamcorper tortor magna commodo mauris enim suscipit consequat, bibendum nec nascetur lectus praesent facilisi. Montes ornare ad orci lobortis bibendum, neque diam eget scelerisque cursus condimentum, donec semper mauris accumsan."
+        }
     }
 }
